Add cgo to build tags
