+++ journalbeat/reader/fields.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//+build linux
+//+build linux,cgo
 
 package reader
